Add a network path monitor and dump ifaces.
This is in tangent with the following commit. Where it seemed that iOS wanted us to specify the host interface sometimes, especially when we're on a Self AP. Similar to how Android behaves. This turned out to be wrong though because the same network that worked once would work twice. So, this is just me dumping interfaces.