Don't stop DNS group when asked.
Previously we were trying to be friendly to the operating system by shutting down things we didn't need when we'd lose focus (and there for be stopped)

Only, this seems to cause problems. Things are working much better for me if I just leave this running all the time.

It may be time for a suspend/resume hook.